move functions to ceneo check
